third_party/blink/web_tests/fast/forms/select/customizable-select/nested-select-console-message-expected.txt

CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select><select>. Please add the missing </select> end tag.
option
